Revert "Refactore"
