+++ src/components/Header/header.jsx
@@ -1,12 +1,11 @@
-import s from './header.module.scss';
 import {BrowserRouter, Routes, Route, Link} from 'react-router-dom';
 import LogoLivro from '../../assets/logoLivro.png';
-
 import Inicio from '../../pages/Inicio/inicio';
 import LivrosDoados from '../../pages/Doados/doados'
 import QueroDoar from '../../pages/QueroDoar/queroDoar';
-import Info from '../../pages/Info/info';
+import Clientes from '../../pages/Clientes/clientes';
 import Lupa from '../../assets/search.png';
+import s from './header.module.scss';
 
 export default function header() {
     return (
@@ -35,7 +34,7 @@ export default function header() {
                 <Route path='/' element={<Inicio/>}/>
                 <Route path='/doados' element={<LivrosDoados/>}/>
                 <Route path='/QueroDoar' element={<QueroDoar/>}/>
-                <Route path='/clientes' element={<Info/>} />
+                <Route path='/clientes' element={<Clientes/>} />
             </Routes>
         </BrowserRouter>
     )
