+++ src/pages/Clientes/clientes.module.scss
@@ -1,0 +1,21 @@
+
+.container {
+    // border: 2px solid red ;
+    min-height: 100vh;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+
+    .titulo {
+        // border: 2px solid;
+        font-size: 2vw;
+        margin: 50px 0;
+        @media(max-width: 600px) {
+            font-size: 5vw;
+        }
+    }
+    .containerClientes {
+        // border: 2px solid;
+        width: 60%;
+    }
+}
