+++ src/pages/Doados/doados.module.scss
@@ -29,7 +29,7 @@
         align-items: center;
         overflow: hidden;
         .img {
-            // border: 2px solid;
+            border: 2px solid;
             width: 5vw;
             margin-top: 0.9rem;
         }
