+++ src/pages/Inicio/inicio.module.scss
@@ -33,12 +33,11 @@
         // border: 2px solid;
         border-radius: 15px;
         width: 20vw;
-        height: 50vh;
+        height: 45vh;
         display: flex;
         align-items: center;
         flex-direction: column;
         background-color: #0157944D;
-        cursor: pointer;
         .imgCards {
             // border: 2px solid;
             width: 5vw;
